refactor(command): extract provider transfer
